comments : move around some github specific stuff
