+++ Makefile
@@ -32,6 +32,7 @@ ghcli_SRCS			=	src/ghcli.c			\
 					src/json_util.c			\
 					src/config.c			\
 					src/github/repos.c		\
+					src/github/comments.c		\
 					thirdparty/sn/sn.c		\
 					thirdparty/pdjson/pdjson.c
